+++ src/main/java/com/mylog/common/annotations/Password.java
@@ -1,9 +1,10 @@
-package com.mylog.common.validation;
+package com.mylog.common.annotations;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
+import com.mylog.common.validation.PasswordValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 import java.lang.annotation.Documented;
