+++ src/main/java/com/mylog/common/exception/GlobalExceptionHandler.java
@@ -1,7 +1,6 @@
 package com.mylog.common.exception;
 
 import com.mylog.common.response.ErrorResponse;
-import lombok.NonNull;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
