+++ src/main/java/com/mylog/common/security/SecurityConfig.java
@@ -5,6 +5,7 @@ import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -46,7 +47,9 @@ public class SecurityConfig {
         .authorizeHttpRequests(
             auth ->
                 auth.requestMatchers(WHITELIST)
-                    .permitAll() // 해당 url 허용
+                    .permitAll()
+                    .requestMatchers(HttpMethod.GET, "/api/articles/**")
+                    .permitAll()
                     .anyRequest()
                     .authenticated() // 나머지 접근 방지
             )
